split out panel state
